Modify create_job to send up the content_type to the emotion as a service api
Without this change it sends up a blank content type, which is not ideal because
the processing code uses this to determine if we need to transcode to the video
to MP4.

This will be a slight performance improvement because we will no longer be transcoding
to MP4 when we don't need to.