Add a create_entry() method
Sometimes we want to "park" media in EaaS but don't need to process
it, and since the REST API allows for creation of entries it looks
like it will be useful to expose that via the client bindings.  I
added optional annotations to the method since we almost always add
annotations to the entries.